requirements.txt
Flask==3.0.2
gunicorn==21.2.0
numpy==1.26.4
pillow==10.2.0
tensorflow==2.15.0
waitress==3.0.0
